requirements.txt
openai>=1.54.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
python-frontmatter>=1.1.0
python-dotenv>=1.0.1
duckduckgo-search>=6.3.0
